crowbar_framework:fix filtered nodes in apply_role
As a side effect of introducing the experimental options for unchanged
and unready nodes into apply_role, there could be some cases in which
even if the nodes where removed from the list to be deployed we still
deployed them. For that issue, a check is done before adding the nodes
into the batches to check that indeed those nodes are going to be
deployed.

There was also a wrong comparision being done of the old_nodes vs
new_nodes in which we used the filtered list of nodes, but we want to
use the original node list without unfiltered to make sure that the
deployments didnt change. For this,the unfiltered list of nodes in the
new deployment is regenerated and used for the role comparision.