skip unchanged nodes from chef run
By using skip_batch_for_node? we can avoid running chef on nodes that do
not change role and if there is no attribute change for the barclamp.

Does not skip nodes that have changed roles, all nodes if
attributes have changed and when deploying the barclamp for the first
time.

This commit is a set of shared methods to analyze changes in roles
and attributes.

Requires: #1401
(cherry picked from commit 6c4451439744e3ed86cc64a966ed08a612c2da10)