don't retry with installed state during failing crowbar_join (bsc#1063533)
If chef-client fails twice during crowbar_join, the logic is to try
running it a third time in "installed" state with client.pem removed,
to pretend we just come straight from a fresh install.  This is very
old logic originally written by Victor, and we don't know of any
scenario where this actually helps.  In contrast, we do know that it
is quite easy for chef-client to fail on nodes which successfully
installed a long time ago and are just running crowbar_join again on
boot up, e.g. after zypper up + reboot.  And in this case this logic
removes a perfectly good client.pem forever:

    https://bugzilla.suse.com/show_bug.cgi?id=1063533

Worse, pretending that a previously installed node is being freshly
installed could potentially lead to serious data loss, so it's safer
to just bail at this point and let the operator fix it.

(cherry picked from commit 5d7974dc2909312070153831882a8639a148eb38)