service_object: refactor apply_role
Refactor apply_role to be split into 4 different methods
that sync with the four different things that we were doing
previously in one big method.

There is not too manya changes to the actual methods apart
from removing a coouple of not needed attributes where it
makes sense because they are available already in the method

Uses exceptions to catch and return the proper status and
error messages to crowbar.