crowbar: Fix race on node reinstall
Looking at logs, it turned out that we were rebooting the node before
chef-client was run on the admin server. This is due to the fact that we
run a script (single_chef_client) that will detach to the background and
then launch a script (looper_chef_client) that will touch a file
(/var/run/crowbar/chef-client.run) to trigger a chef run. And this went
slower than the rails code.

So add an explicit sleep for 2 seconds in the rails code. It's kind of
ugly, but on the other hand, we know that chef-client will run and it
will take some time; so waiting a couple of seconds won't actually delay
anything.

(cherry picked from commit 285e8bc744be7a9bd5568fe7b12a5923a8d4bc1a)