utils: Load databag item once per run
For checking the disallow_restart flags, we load a databag item. At the
moment, it was loaded every time we were checkign the flag for a
service.

By caching it, we get some small performance improvement, but we also
get consistency by making sure that the flag doesn't get changed in the
middle of a chef run.