crowbar : priorice provisioner barclamp execution
After restart a node, Keystone requires the fernet keys and
they have to be imported from the founder node. To do that
is necessary to do a rsync, and to do that the ssh keys are
required. However in the current barclamp sort, the provisioner
is executed at the end of the list. So, In order to have the
ssh keys from the begining to this fix priorize provisioner
barclamp to be executed from the begining.