network: Remove slave config when preparing a bond
As part of the bond setup we're taking the slave interfaces down. To
avoid that a later network restart reassigns an IP address to the slave
interface(s) we also need to remove the configuration files for the
interfaces as early as possible.

(cherry picked from commit 1ae06dedf3b13c3f38301aa5b8dac7bb49f6a03b)