network: Fix re-config wait loop (bsc#1064813)
To get the IP of the provisioner we need to read a Data Bag item from
the server. Trying to read directly before the wait loop for the network
to come back doesn't make sense, as we might have connectivity issues at
that point. Move it back to the top of the recipe where it was before
commit 8d9265fb0.