provisioner: Restore some small Hyper-V files
It's fine to remove the large binary files, but removing small text
files makes no sense unless we drop all of them (there are more) and
unless we really remove Hyper-V support from all recipes (that was not
done).

This makes it slightly easier to restore things in case someone wants to
restore that later.