wrap lines from Chef logs in <pre> element
This avoids the horrendously ugly UI mess which we've somehow
tolerated for a few hours, by ensuring the lines are in a monospaced
font, and the size of the block capped to a maximum, with scroll bars
if required.

The background colour of the <pre> element needs to be set to
something otherwise it will be white which looks weird, so choose a
colour which is very slightly lighter than its surrounding alert's
background colour.