crowbar: Ensure epmd.socket is configured correctly
With new erlang and rabbitmq-server packages, rabbitmq now depends on
the system-wide epmd service, which will only listen on IP addresses
defined for epmd.socket.

We need epmd to listen on the admin IP address (the one that resolves
the FQDN), so configure things accordingly.

While install-suse-cloud will already do this setup correctly, we also
change the crowbar cookbook to ensure the config stays up-to-date.