crowbar-hacks: Skip activation of HPE UEFI entries
HPE servers have some standard UEFI boot entries which are disabled
by default. After crowbar activates them, they are re-created by the
firmware with disabled state. Crowbar activates these duplicates and
the whole cycle repeats after each reboot. This can lead to very long
lists of boot entries which can even cause the machine to fail booting.

The workaround for this problem is a list of entry names which are
skipped during the activation loop. Currently the list includes only
the HPE specific entries but can be extended in the future if needed.