provisioner: Don't use net_bootp for PXE-booted EFI-enabled grub
net_bootp would configure all network interfaces with DHCP (which is
more than what we want, and may be slow).

However, it turns out that the efinet driver in grub2 will already
know the configuration configure of the network interface that was used
for PXE booting (since grub is parsing the DHCP ACK that the EFI
firmware received).

See grub_efi_net_config_real and grub_machine_get_bootlocation from:
http://git.savannah.gnu.org/cgit/grub.git/tree/grub-core/net/drivers/efi/efinet.c
http://git.savannah.gnu.org/cgit/grub.git/tree/grub-core/kern/efi/init.c
and for the patch introducing this:
http://git.savannah.gnu.org/cgit/grub.git/commit/?id=cae730b4526f250a3d5073c00911e35b7abdd7d7

Co-authored-by: Stefan Seyfried <stefan.seyfried@sap.com>