provisioner: Let grub2 provide the MAC address for BOOTIF in UEFI mode
In UEFI PXE boot mode, grub will know from which interface we're
PXE-booting from, and can provide automatically the correct MAC address.

This will be fine, because the PXE config is fine (since we configure
DHCP to only boot from the interfaces that we want).

This solves issues when, for instance, the discovery image boots from
one interface that turns out to not be in the admin conduit after
network.json resolution.

Generally speaking, this also makes sense because we simply rely on
"reality" (since grub will put the right value), and not on something
that could be possibly wrong due to outdated chef data.

Co-authored-by: Stefan Seyfried <stefan.seyfried@sap.com>