crowbar: clean restart flags extended
This commit extends the functionality of #1307 and #164. These commits allow to the final user to disable the restart of services when the chef client runs. If it is active for any cookbook and one of its services need to clean the flag is raised. To clean this flag there is the command "clear_restart" that the user will run when after the manual restart. In this command the service name have to be specified. However is interesting to restart all services of the one cookbook or restart all services in a node.

This commit allows to specify the name of the cookbook, cleaning all the service of this cookbook, also allows to don't specify any cookbook that will clean all flags in all cookbooks.

(cherry picked from commit ea3fe89f7062ad82256cd7963f580225675f5bb0)