crowbar: clean restart flags extended
This commit extends the functionality of 690e9bb. This commit allows to the final user
to disable the restart of services when the chef client runs. If it is active for any
cookbook and one of its services need to clean the flag is raised. To clean this flag
the command "clear_restart" could be run by user after the manual restart.

The common usage is to specify the name of the service. However this modification allows
to restart all services of the one cookbook (specifying the name of the cookbook instead
of the service name) or to restart all services in a node (if the name of service is
omitted).

Additionally, the alias could be used instead of the name of the node.

(cherry picked from commit e795aba4375a1fc95fcd891ffe2a82fe85033ce5)