Dangerfile: Extend subject check to 70 chars
Rationale: we use them for rpm changelogs that are visible
to customers, so we want to have a little bit more detail in
there like a crowbar module prefix and an bugzilla id at the end.

The commonly accepted industry standard of 50 chars per commit subject
is therefore extended to 70 chars.