upgrade: Remove pre-upgrade constraints from existing locations
pre-upgrade constraints ensure that the service is not running at the
node with active pre-upgrade attribute. This is necessary during the upgrade
when we need to control which nodes can run services and which cannot.

So we must do a cleanup before we start the upgrade, otherwise all services would
stop at all nodes once we set pre-upgrade attribute to all nodes (which is the
initial step of the upgrade).

(cherry picked from commit e9d93dcdad1da698060829e874bdf1f1f7492211)