upgrade: fix horizon deprecated config
As during the upgrade we first update the packages, and then update
the configuration (via chef) when upgrading the node packages, any
horizon related package will fail to install due to those packages
%postun firing a manage.py command which loads our custom
_100_local_settings.py which contains a deprecated config item.
As by the time those packages are installed, django has already been
updated, it will fail to finish the install.

This patch will replace the outdated configuration in the config
prior to upgrading the packages so the package install can finish
correctly. The configuration will be upgraded by chef afterwards
so it will end up being correct no matter what.