crowbar: add validator for n nodes or cluster
There's some situation where the proposal validators need to check that
a number of nodes are present. This validator is already present. But if
a cluster is assigned to the proposal, then the cluster it is counted as
a single instance.

A new validator 'validate_multiple_for_role_or_cluster' has beed
added. It does check that at least 2 nodes are present or a cluster is
defined.

(cherry picked from commit b4b65638bde309d9eaa1e4eb57af782f61807c57)