crowbar: Fix Node.get_network_by_type for overridden attributes
We were looking for attributes overriding the network proposal only on
the node role, but this is not consistent with what the cookbook does
(it looks at attributes on the node itself), and customer using this
feature are more likely to edit the node directly.

So reflect that by looking at the node attributes also. This is slightly
tricky, as the merge of node attributes and node role attributes occur
on a chef run, and in the rails app, this can be out of sync. So we do
our own merge to always see the real values.