upgrade: Do not upgrade compute-related packages to Pike
It does not seem to be necessary to upgrade openstack packages
to from Newton to Pike version for them to work with Pike-based control plane.

Restart of current services (after control plane upgrade) is enough.

Note: these specific packages will be upgraded at some point,
but together with all packages on the node.