upgrade: Allow running schema migrations on upgrade
This is basically a revert of 73357d9318084e73abc0a512eaeb5ba8c951c7c6.

The code made sense in 6-7 upgrade, as we were creating new database and the migration
was explicitely called after data import.

For 7-8 upgrade, we are not changing crowbar database so we need to run the migrations
after package upgrade.