upgrade: Make sure services are stopped when corosync is shut down
Right before packages on a node should be upgraded, all OpenStack related
services should be stopped (so that package upgrade does not trigger schema migration).

For traditional HA case, this is achieved by shutting down corosync which in turn shuts
down all services, But when clone_stateless_services is set to false, pacemaker does not manage
all services and they need to be stopped explicitely.