upgrade: Shutdown remaining OpenStack related services in the cluster
After first node has been upgraded, all services on other nodes of the
cluster need to be shut down (they will be running on upgraded node only).

With traditional HA, this is achieved when all pacemaker resources are deleted.
However for the case when services are not managed by pacemaker, they should
be shut down explicitely.