upgrade: disable network agents during upgrade
As a workaround for https://bugs.launchpad.net/neutron/+bug/1780453 we
are now setting the "ADMIN_STATE_UP" attribute to "down" before
upgrading a (network)-node. The attribute is reset to "up" after the
node was successfully upgrade and all services are running again. This
should e.g. avoid neutron/neutron-dhcp-agent to try to bring up e.g. a
dhcp port on a node where the Layer 2 agent (openvswitch-/linuxbridge-agent)
is not fully up yet.