upgrade: Port openstack db backup to mysql
This was still relying on postgresql being the database backend for
openstack. It went unnoticed for so long because the default values for
the postgresql backend related chef attributes pointed the upgrade
procedure to "localhost". So it created a dump of the admin node's
database.

(cherry picked from commit 4c95e5b48cfeb9aa16f5cb4ded386d182ae6850b)