upgrade: Allow to shutdown postgresql and drbd resources
posgresql and drbd resources no longer need to be excluded from
the list of services that are going to be shut down.

Neither of those is supported in SOC8 and neither of them should
be used by SOC7 when the upgrade is started (if there was PostgreSQL
setup initially, it had to be migrated to MariaDB).