upgrade: Exclude removed nodes from upgrade
When some node is removed from crowbar (e.g. using "forget" operation) and
not all proposals are saved, this node was still included in some parts of
the upgrade process causing errors when trying to access node attributes.

Additional filtering was added to make sure we don't try to upgrade
non-existent nodes.