crowbar: move disallow chef restarts out of experimental
Looks like this feature has been tested extensively on prod envs
with little to no failures so its time to move it out of the
experimental umbrella, make a normal config for it and document
it properly.

Moves the config settings from the experimental to the normal config
Changes the calls to experimental settings to point to the new file

Removed experimental.yml because it won't be necessary.

(cherry picked from commit 904aa5754c4999ef7ca19b225f6d5accf0997190)