upgrade: Do not start chef-client service from recipe
Originally chef-client was started when running crowbar_join after
the node upgrade. But we still need to save some node data after
this point (in finalize_node_upgrade) and there could be a chance that
chef-client is already running at the node. In such case, our node.save
would not have an effect and running chef-client saves the old reversion
of a node.

To prevent this, do not start chef-client service from that crowbar_join
but only explicitely from finalize_node_upgrade after the node data were
saved.