provisioner: validate format of additional ssh keys section
The chef recipes crash with a very weird error message when
the additional ssh keys section is malformatted, which can easily
be the case given that ssh keys are usually copy&pasted. Provide
some useful hints for troubleshooting formatting errors and enforce
that we have at least 3 sections because the chef recipes
are doing unification based on the "comment" field.