Add ping support for IPv6
Currently the SLE12 iptuils package is using iputils 20121221 which
has separate ping and ping6 binaries. Which is fine, but the ping
command only works with ipv4, even if given a FQDN the qaulifies to
ipv6.

SLE15's iputils is using 20161105, which has both ipv4 and ipv6 in
support in the ping command.

This change modifies the check to test with ping6 if ping fails, so
the check will work on legecy systems.

The other alternative is to update SLE12's iputils package which
may be a better alternative, but this change bypasses the problem.