upgrade: Introduce parallelism into the upgrade of compute nodes
Originally we upgraded compute nodes one by one, for non-disruptive mode.

But it might be possible to live-evacuate more than one compute node at a time.
In such case we can upgrade all the nodes that currently do not host
running instance. However, let's not try to be too clever and let the selection
of nodes subset that should be upgraded in parallel up to user. This means that
the original option to selectively upgrade one compute node gets updated
and supports list of nodes, separated by comma, space or semicolon.