upgrade: Add an initial check if there's anything running on a host
This is necessary after some failure and retry.
When the compute node upgrade fails e.g. during OS installation and
user retries the step, first think that crowbar does is to live-evacuate
the host that has failed the upgrade. However the script for live-evacuation
could fail because nova-compute service might be stopped on the node already.

So we add the initial check that makes sure there are no running instances
on the host (they should not be, otherwise the OS upgrade would not start).

(cherry picked from commit 9e66773db2a339c9d2801ed2f686c1a26da5eed2)