upgrade: If pacemaker is used to manage the services, don't use systemctl
Usage of "systemctl stop" for shutting down the services should be limited
to the case when pacemaker is not managing those services, i.e. clone_stateless_services
is set to false. However, we're changing the value of this variable during the upgrade
so we must save the original one first.