Ipv6 support in the crowbar_validator
The crowbar validator uses the barclamp schema files to determine
if the current values are correct. When attempting to validate the
network.json file for the network barclamp that validator doesn't
understand ipv6 addresses.

This patch adds some extra regexes to the IpAddress validation to
understand not just IPv4 addresses but also IPv6 and CIDR netmasks.

This is one step closer to getting IPv6 support in crowbar.