Sanitise the IP address strings returned from resolv
The resolv module will return the IPs as strings exactly as it
finds them. When dealing with ipv6, this can be a problem, as
IPv6 allows shortcuts when writing IPs.
This can make However when comparing IPs as strings fail, say in
the ip_configured method, when comparing IPs in the socket and what's
returned by resolv.

This patch santises the string by using the IpAddress to_s method which
uses the same IPv6 address format used by the Socket.