apache: keep vhost config when disabling apache site
On SUSE the apache_site definition used to just remove the vhost config
file when disabling a certain vhost config. This commit changes the code
so that the vhost config file will be renamed to <vhost.conf>.disabled.
That way apache will ignore it but the config is still present.

The rationale for this is that during the upgrade we want to create a
"disabled" configuration for a vhost (keystone in this case), that we
need enable again manually later (outside of chef) by just renaming the
file and reloading apache.

(cherry picked from commit a7dde6643922b5a8af066586dde540e617561216)