salt: Make the barclamp visible in the UI and adjust order
To be able to deploy salt on the admin node (and use it via salt-ssh),
the baclamp must be visible. That's done now.
Also move the barclamp in the list down (given that it is not that
important and more an optional feature).
Also change the "proposal_schema_version" to 3 which seems to be the
default in other barclamps.