apache2: enable sslsessioncache
The if conditionals in this files had a typo which made then
always disabled. We need to list the module name like documented.

Merge it more with the package-provided default config file.

Also update cipher list away from a very insecure default and
use the openssl provided SUSE_DEFAULT cipher list so that we don't
ever have to update this again.