upgrade: Delay status switch after upgrade ends
Previously the "next upgrade" status file was used as soon as "current"
upgrade was finished. This sometimes caused problems when "final" status
was missed because it was already covered with new "not started yet" status.

With this change, "old" status will be returned for 24hrs after upgrade was
finished. After that subsequent calls to status API would return "new" status.