Fixup uid/gids if they don't match (SCRD-8014)
To fix the ones on disk from rpm packages by reinstalling the package in
question with --force.

The rpm package scripts in question just pick a free id. It can not be
ensured that creating the user/groups with specific ids is done before
the packages are installed as crowbar_register may be run on an existing
system that has one or more of the packages already installed. Also
there is the update case, though less likely by now as this code already
existed in the last release.

Note that this does not fix up any on disk uid/gids for files that may
have been created by crowbar (which should be fixed by the following
run) or manually.