barclamp: Fix setting MTU on networks using a bridge
When add_bridge is set to true, we set the MTU on the bridge interface,
but nothing sets it on the slave interface, so it doesn't work.

So whenever we set the MTU on a bridge, also set it on the slave
interface.

(cherry picked from commit 651d40940d7f80e7a078563dbb3abfeedc47723e)