ipv6: append the adminip to the cmdline of the discovery image
When the sleshammer discovery image boots on nodes in an IPv6
environment there is no sane way of determining the admin ip.

In IPv4 we can grab it from the SERVERID produced by wicked's
dhcpd4 binary. Except in IPv6 the serverid seems to the be DHCP
DUID.

This patch adds an 'adminip=..' option to the commandline for
the discovery image. The start-up.sh script and sleshammer image
has been modified seperately.