gather_logs: Make it a bit useful again
We were listing nodes but on stdout (so that was not in the archive that
is generated), and the commands we were running to gather proposals and
roles were not working (since the move to crowbarctl, at least).

Fix this by using the right commands, and also storing the data for
nodes, proposals and roles in json files in the archive.

(cherry picked from commit b34c6ae4595dea14a610f880aa86d434aa7681a7)