transition.sh: Do not read /etc/crowbar.install.key
I could not find any use of this script... However, given that it relies
on "knife node list" and that it's designed to be a script run by user
(as opposed to run by crowbar), we can assume it is only run on the
admin server as root.

The script only uses crowbarctl and we know we have /etc/crowbarrc, so
there's no need for reading /etc/crowbar.install.key or anything close
to it.

(cherry picked from commit 822cc1ce4745729ecd5c5512f09ecdd00a5aafc2)