crowbar: Allow hardware-installing -> discovering transition (noref)
Strictly speaking, "hardware-installing" -> "discovering" should not be
possible. However, if the node is "hardware-installing" and it reboots
before chef-client on the node completes a run that will make sure it
inherits all the right attributes, then dhcp will still be configured
with the discovery IP address (due to missing chef attributes) and the
discovery process will start again.

It could be argued it's a race, or a flaw in how we configure dhcp. It
would be worth fixing, but for now, this will do.

(cherry picked from commit 8e5f582519ebc7f3275514b5fe212dfc18bd1879)