crowbar: Use /etc/crowbarrc as unique source of admin credentials
The problem of using attributes of the crowbar proposal for admin
credentials for the rails app is that these credentials are available
through chef, and therefore any node in the deployment can get them and
could operate destructive operations on the hardware (like reinstalling
a node).

Instead of using the attributes of the crowbar proposal to configure
admin credentials, we now use /etc/crowbarrc on the admin server as
source of this information. Note that we use the "default" section of
that configuration file.

We still need chef-client on the admin server to be able to update some
options in /etc/crowbarrc (server, verify_ssl), but it will not touch
the username and password anymore.

As a side effect, this also means we can't have multiple users for admin
access anymore. This was however not used outside of the machine-install
user which is now not needed anymore.

This is related to bsc#1117080, SCRD-7537, CVE-2018-17954.