crowbar: Move crowbarrc mgmt into crowbar cookbook (SCRD-8330)
In the case of a restoration from backup, /etc/crowbarrc will already
have been modified by the last installation. During the restore, the
install-suse-chef.sh script runs chef on the admin node, but only with a
select set of roles which at certain points does not include the
provisioner role. The result is that from crowbar's point of view at
that point in time this is a new installation and the proposal values,
including whether to use SSL, are set to default values. However, the
crowbarrc file may already have been modified to use SSL. In this case,
trying to use crowbarctl will fail because the crowbar server is not yet
configured for SSL.

This patch addresses the problem by moving the management of the
crowbarrc file into the crowbar cookbook for the admin node. This way,
if only the crowbar role is included, it can still properly manage the
crowbarrc file so that the server and verify_ssl parameters are correct
for the stage it is running in.