network: Rework the vlan port replugging code (bsc#1120657)
The current workaround for plugging a network device (e.g. eth0) into an
ovsbridge (e.g.  br-fixed) when a VLAN device of that (e.g. eth0.42) is
already plugged into a different  ovs-bridge (e.g. br-public) produces
an error message on the logs. This commit reworks the replugging
workaround to avoid that error, by first unplugging the VLAN ports and
replugging them after the base port is attached.