network: Allow locking down the network config for nodes (bsc#1120657)
This change provides the possibilty to lock down the network
configuration of specific nodes. It implemented by simply skipping the
whole network cookbook whe the nodes attribute
'node["network"]["config_frozen"]' is set to 'true'. In that case
anychanges to the network barclamp, or any other change that would cause
the network to be reconfigured (e.g. allocation of IPs, enablement of a
network, addition or removal of network devices, ...) will just have no
effect on that node.