IPv6: ntp.conf requires netmask in non CIDR format
The ntp.conf used in crowbar wont work with IPv6 as the netmask
in the ntp conf file need to be in the full form not CIDR.

This patch adds a new method, cidr_to_subnet to the Barclamp library's
Network class. Which is then used to send the correct string to the
ntp.conf.erb template.

Allowing the crowbar nodes being booted via the sleshammer discovery
image to sync with the admin node.