travis: pin sexp_processor to 4.12.0
we use brakeman during our travis run. It has a dependency
sexp_processor (deep buried). This gem got a new commit[1] which is not
compatible with brakeman.

The best case would be to fix brakeman and backport it to v2.6.3 (also
pinned) which had a last commit in 2014.  So pinning the dependency is
the easier fix

1 - https://github.com/seattlerb/sexp_processor/commit/ce284487f057203360c41b14d2b25f8c5453fbb9