Delete controller on ovs bridge for admin net at boot (bsc#1141124)
On a clean shutdown wicked deletes any ovs bridges it manages. On the next boot
wicked will recreate them without a controller, in standalone mode and they will
forward traffic immediately. When the neutron ovs agent starts, it will add a
controller, set fail mode to secure and add new flows.

When the node is not shutdown cleanly, the bridges won't be deleted from the
ovs database. On the next boot, ovs (and not wicked) restores the bridges from
the database along with the controller and fail mode settings. The pre-up script
runs del-fail-mode which puts the bridge into standalone mode, but with a
controller still configured, the bridge will continue to try to connect the it.
It will not start forwarding traffic until it enters fail-open mode after 15s
(3x inactivity probe). When the neutron ovs agent starts, the bridge will
connect to it immediatelty and drop out of fail-open mode and stop forwarding.
If the neutron ovs agent requires the bridge to reach the rabbitmq server, then
it will get stuck and won't be able to reach the point where it installs the new
flows to make the bridge start forwarding again.

This patch fixes that by deleting the controller from bridges that provide the
admin network in the pre-up script.