Public ips for dns nodes when designate integration is in use (SOC-9635)
When crowbar's DNS is set to intergrate with designate we need the DNS
servers to be listening on the public network so tennent users of
desigante can access the zones they create via the API.

If designate intergration is enabled, this patch allocates a public ip
for each DNS node. Each node's bind9 is also setup to listen on both the
public and admin ips.